+++ src/virtdisk_bindings.rs
@@ -20,7 +20,7 @@ extern "C" {
         virtualDiskAccessMask: VirtualDiskAccessMask,
         securityDescriptor: *const SecurityDescriptor,
         flags: u32, // create_virtual_disk::Flag
-        providerSpecificFlags: u64,
+        providerSpecificFlags: u32,
         parameters: *const create_virtual_disk::Parameters,
         overlapped: *const Overlapped,
         handle: *mut Handle,
@@ -30,7 +30,7 @@ extern "C" {
         virtualDiskHandle: Handle,
         securityDescriptor: *const SecurityDescriptor,
         flags: u32, // attach_virtual_disk::Flag
-        providerSpecificFlags: u64,
+        providerSpecificFlags: u32,
         parameters: *const attach_virtual_disk::Parameters,
         overlapped: *const Overlapped,
     ) -> DWord;
@@ -38,33 +38,33 @@ extern "C" {
     pub fn DetachVirtualDisk(
         virtualDiskHandle: Handle,
         flags: u32, // detach_virtual_disk::Flag
-        providerSpecificFlags: u64,
+        providerSpecificFlags: u32,
     ) -> DWord;
 
     pub fn GetVirtualDiskPhysicalPath(
         virtualDiskHandle: Handle,
-        diskPathSizeInBytes: *const u64,
+        diskPathSizeInBytes: *const u32,
         diskPath: PWStr,
     ) -> DWord;
 
     pub fn GetAllAttachedVirtualDiskPhysicalPaths(
-        pathsBufferSizeInBytes: *mut u64,
+        pathsBufferSizeInBytes: *mut u32,
         pathsBuffer: PWStr,
     ) -> DWord;
 
     pub fn GetStorageDependencyInformation(
         objectHandle: Handle,
         flags: u32, // storage_dependency::GetFlag
-        storageDependencyInfoSize: u64,
+        storageDependencyInfoSize: u32,
         storageDependencyInfo: *mut storage_dependency::Info,
-        sizeUsed: *mut u64,
+        sizeUsed: *mut u32,
     ) -> DWord;
 
     pub fn GetVirtualDiskInformation(
         virtualDiskHandle: Handle,
-        virtualDiskInfoSize: *const u64,
+        virtualDiskInfoSize: *const u32,
         virtualDiskInfo: *mut get_virtual_disk::Info,
-        sizeUsed: *mut u64,
+        sizeUsed: *mut u32,
     ) -> DWord;
 
     pub fn SetVirtualDiskInformation(
@@ -74,21 +74,21 @@ extern "C" {
 
     pub fn EnumerateVirtualDiskMetadata(
         virtualDiskHandle: Handle,
-        numberOfItems: *mut u64,
+        numberOfItems: *mut u32,
         items: *mut Guid,
     ) -> DWord;
 
     pub fn GetVirtualDiskMetadata(
         VirtualDiskHandle: Handle,
         item: *const Guid,
-        metaDataSize: *mut u64,
+        metaDataSize: *mut u32,
         metaData: *mut Void,
     ) -> DWord;
 
     pub fn SetVirtualDiskMetadata(
         virtualDiskHandle: Handle,
         item: *const Guid,
-        metaDataSize: u64,
+        metaDataSize: u32,
         metaData: *const Void,
     ) -> DWord;
 
@@ -146,7 +146,7 @@ extern "C" {
         byteLength: u64,
         flags: u32, // query_changes_virtual_disk::Flag
         ranges: *mut query_changes_virtual_disk::Range,
-        rangeCount: *mut u64,
+        rangeCount: *mut u32,
         processedLength: *mut u64,
     ) -> DWord;
